requirements/base.txt
# Django
django>=5.0
django-environ>=0.11
whitenoise>=6.6
pillow>=10.0
python-dateutil>=2.9

# PostgreSQL
psycopg2-binary>=2.9

# Redis
redis>=5.0
celery>=5.3
django-celery-results>=2.5

# Social Auth (OAuth2)
social-auth-app-django>=5.4
social-auth-core>=4.5
djangorestframework-simplejwt>=5.3

# REST API (DRF)
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27
strawberry-graphql[django]>=0.235
strawberry-graphql-django>=0.47

# Field Encryption
django-encrypted-model-fields>=0.6
cryptography>=42.0

# LLM / AI
anthropic>=0.39
langchain-anthropic>=0.3
langchain-core>=0.3
langgraph>=0.2
langchain-community>=0.3
tavily-python>=0.5

# Monitoring & Logging
sentry-sdk[django]>=1.40
